requirements.txt
Flask
transformers
torch
scikit-learn
python-dotenv
langchain_community
langchain_aws
flask_cors
cors
joblib